refactor: move todoList and isLoading state from App.js to TodoContainer.js.
